Continue to simplify and reorganize
